+++ jirkov.vladimir/T2/DataStruct.hpp
@@ -1,10 +1,7 @@
 #ifndef DATASTRUST_HPP
 #define DATASTRUCT_HPP
 #include <string>
-#include <utility>
-#include <iostream>
-#include "ScopeGuard.hpp"
-#include "IOstructures.hpp"
+#include <iosfwd>
 namespace jirkov
 {
   struct DataStruct
@@ -12,12 +9,9 @@ namespace jirkov
     unsigned long long key1;
     unsigned long long key2;
     std::string key3;
+    bool operator>(const DataStruct &left, const DataStruct &rigth);
+    bool operator<(const DataStruct &left, const DataStruct &rigth);
   };
-
-  std::istream & operator>>(std::istream &in, DataStruct &value);
-  std::ostream & operator<<(std::ostream & out, const DataStruct &value);
-  bool operator>(const DataStruct &left, const DataStruct &right);
-  bool operator<(const DataStruct &left, const DataStruct &rigth);
 }
 
 #endif
