+++ jirkov.vladimir/T2/DelimiterIO.hpp
@@ -1,0 +1,14 @@
+#ifndef DELIMITER_IO_HPP
+#define DELIMITER_IO_HPP
+
+namespace jirkov
+{
+  struct DelimiterIO
+  {
+    char exp;
+  };
+}
+
+#endif
+
+
