+++ jirkov.vladimir/T2/IOstructures.cpp
@@ -1,6 +1,10 @@
 #include "IOstructures.hpp"
 #include <iostream>
-
+#include "DelimiterIO.hpp"
+#include "UllOctIO.hpp"
+#include "UllLitIO.hpp"
+#include "StringIO.hpp"
+#include "DataStruct.hpp"
 std::istream & jirkov::operator>>(std::istream & in, DelimiterIO && value)
 {
   std::istream::sentry guard(in);
@@ -19,23 +23,18 @@ std::istream & jirkov::operator>>(std::istream & in, DelimiterIO && value)
 
 std::istream & jirkov::operator>>(std::istream & in, UllLitIO && value)
 {
-  std::istream::sentry sentry(in);
-  if (!sentry)
+  std::istream::sentry guard(in);
+  if (!guard)
   {
     return in;
   }
-  in >> value.ref >> DelimiterIO{'u'} >> DelimiterIO{'l'} >> DelimiterIO{'l'};
-  if (!in)
-  {
-    in.setstate(std::ios::failbit);
-  }
-  return in;
+  return in >> value.ref >> DelimiterIO{'u'} >> DelimiterIO{'l'} >> DelimiterIO{'l'};
 }
 
 std::istream & jirkov::operator>>(std::istream & in, UllOctIO && value)
 {
-  std::istream::sentry sentry(in);
-  if (!sentry)
+  std::istream::sentry guard(in);
+  if (!guard)
   {
     return in;
   }
