+++ jirkov.vladimir/T2/IOstructures.hpp
@@ -2,29 +2,12 @@
 #define IOSTRUCTURES_HPP
 #include <string>
 #include <iosfwd>
-
+#include "ullOctIO.hpp"
+#include "UllLitIO.hpp"
+#include "StringIO.hpp"
+#include "DelimiterIO.hpp"
 namespace jirkov
 {
-  struct DelimiterIO
-  {
-    char exp;
-  };
-
-  struct UllLitIO
-  {
-    unsigned long long &ref;
-  };
-
-  struct UllOctIO
-  {
-    unsigned long long &ref;
-  };
-
-  struct StringIO
-  {
-    std::string &str;
-  };
-
   std::istream & operator>>(std::istream & in, DelimiterIO && value);
   std::istream & operator>>(std::istream & in, UllLitIO && value);
   std::istream & operator>>(std::istream & in, UllOctIO && value);
