+++ jirkov.vladimir/T2/StringIO.hpp
@@ -1,0 +1,13 @@
+#ifndef STRING_IO_HPP
+#define STRING_IO_HPP
+#include <string>
+
+namespace jirkov
+{
+  struct StringIO
+  {
+    std::string &str;
+  };
+}
+
+#endif
