+++ jirkov.vladimir/T2/UllLitIO.hpp
@@ -1,0 +1,12 @@
+#ifndef ULLLIT_IO_HPP
+#define ULLLIT_IO_HPP
+
+namespace jirkov
+{
+  struct UllLitIO
+  {
+    unsigned long long &ref;
+  };
+}
+
+#endif
