+++ jirkov.vladimir/T2/UllOctIO.hpp
@@ -1,0 +1,12 @@
+#ifndef ULLLITIO_HPP
+#define ULLLITIO_HPP
+
+namespace jirkov
+{
+  struct UllOctIO
+  {
+    unsigned long long &ref;
+  };
+}
+
+#endif
